requirements.txt
ipython
django
django-extensions
dj-database-url
Pillow
djangorestframework
requests
ipdb
channels
asgi_redis
requests
daphne
circus
psycopg2
django-bootstrap-form==3.2.1
imagehash
python-dotenv
django-cors-headers
pattern-observer
